+++ janitor-api/src/main/java/com/eischet/janitor/api/modules/DiscoverableModules.java
@@ -1,0 +1,13 @@
+package com.eischet.janitor.api.modules;
+
+import org.jetbrains.annotations.Unmodifiable;
+
+import java.util.List;
+
+@FunctionalInterface
+public interface DiscoverableModules {
+
+    @Unmodifiable
+    List<JanitorModuleRegistration> getModules();
+
+}